chore: Extracted classes into separate files
